Renamed AllCardsServlet
1) Renamed AllCardsServlet to GerCardsServlet.
2) Added package user in package servlet.
3) Removed jetty-plugin configuration from pom.xml.
