+++ pom.xml
@@ -48,11 +48,6 @@
                 <groupId>org.eclipse.jetty</groupId>
                 <artifactId>jetty-maven-plugin</artifactId>
                 <version>9.4.32.v20200930</version>
-                <configuration>
-                    <httpConnector>
-                        <port>9999</port>
-                    </httpConnector>
-                </configuration>
             </plugin>
 
             <plugin>
