+++ src/main/java/com/greetingcard/entity/Status.java
@@ -51,6 +51,5 @@ public enum Status {
                 ", number=" + number +
                 '}';
     }
-
 }
